+++ docs/en/changes/changes.md
@@ -23,9 +23,10 @@
 * Add JSON format support for the `/debugging/config/dump` status API.
 * Enhance status APIs to support multiple `accept` header values, e.g. `Accept: application/json; charset=utf-8`.
 * Storage: separate `SpanAttachedEventRecord` for SkyWalking trace and Zipkin trace.
-* [Break Change]BanyanDB: Setup new Group policy. 
+* [Break Change]BanyanDB: Setup new Group policy.
 * Bump up commons-beanutils to 1.11.0.
 * Refactor: simplify the `Accept` http header process.
+* [Break Change]Storage: Move `event` from metrics to recods.
 
 #### UI